+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Fixed
 
 - Fix project opening not working as expected in IntelliJ 2023 builds.
+- Fix name validation when renaming symbols.
+
+## Changed
+
+- Do not apply spell checking inspection for RuneScript keywords.
 
 ## [1.6.0] - 2024-08-30